feat(defence): implemented defence for post and get link only for owners #6

+++ routes/goods.js
@@ -3,12 +3,15 @@ const Product = require('../models/product')
 
 const router = Router()
 
+const checkEqual = (a, b) => a.toString() === b.toString()
+
 router.get('/', async (req, res) => {
     try {
             const goods = await Product.find()
             res.render('goods', {
             isGoods: true,
             title: 'Goods',
+            userId: req.user ? req.user._id : null,
             goods
         })
     } catch (e) {
@@ -16,37 +19,71 @@ router.get('/', async (req, res) => {
     }
 })
 
-router.get('/:id', async (req, res) => {
-    const product = await Product.findById(req.params.id)
-    res.render('product', {
-        product
-    })
+router.get('/:id/more', async (req, res) => {
+    try {
+        const product = await Product.findById(req.params.id)
+        res.render('product', {
+            product
+        })
+    } catch (e) {
+        console.log(e)
+    }
 })
 
 router.get('/:id/edit', async (req, res) => {
-    const {allow} = req.query
-    if(allow !== 'true') return res.redirect('/')
+    try {
+        const {allow} = req.query
 
-    const product = await Product.findById(req.params.id)
-    res.render('edit', {
-        product
-    })
+        if(allow !== 'true') return res.redirect('/')
+
+        const product = await Product.findById(req.params.id)
+
+        if(!checkEqual(req.user._id, product.userId)) {
+            return res.redirect('/goods')
+        }
+
+        res.render('edit', {
+            product,
+        })
+    } catch (e) {
+        console.log(e)
+    }
 })
 
 router.post('/edit', async (req, res) => {
-    await Product.findByIdAndUpdate(req.body.id, req.body)
-    res.redirect('/goods')
+    try {
+        const product = await Product.findById(req.body.id)
+
+        if(!checkEqual(req.user._id, product.userId)) {
+            return res.redirect('/goods')
+        }
+
+        await Product.findByIdAndUpdate(req.body.id, req.body)
+        res.redirect('/goods')
+    } catch (e) {
+        console.log(e)
+    }
 })
  
 router.post('/delete', async (req, res) => {
-    const cartItems = await req.user.cart.items
-    const idx = cartItems.findIndex(item => item.productId.toString() === req.body.id)
+    try {
+        const product = await Product.findById(req.body.id)
 
-    if(idx < 0) {
-      await Product.deleteOne({_id: req.body.id})
+        if(!checkEqual(req.user._id, product.userId)) {
+            return res.redirect('/goods')
+        }
+
+        const cartItems = await req.user.cart.items
+        const idx = cartItems.findIndex(item => item.productId.toString() === req.body.id)
+
+        if(idx < 0) {
+        await Product.deleteOne({_id: req.body.id})
+        }
+        
+        res.redirect('/goods')
+    } catch (e) {
+        console.log(e)
     }
-    
-    res.redirect('/goods')
 })
 
 module.exports = router
